+++ src/states/gameover.lua
@@ -14,7 +14,7 @@ end
 function GameOver.draw()
 	love.graphics.clear(0, 0, 0)
 	love.graphics.setColor(1, 0, 0)
-	love.graphics.printf("YOU ARE EXPIRED (GAME OVER)", 0, 250, 800, "center")
+	love.graphics.printf("YUCK! YOU ARE INEDIBLE!", 0, 250, 800, "center")
 	love.graphics.setColor(1, 1, 1)
 	love.graphics.printf("Press 'R' to Restart", 0, 300, 800, "center")