+++ Change_Log.md
@@ -77,3 +77,5 @@ Day 9 (12/8/2025):
 
 Day 10 (12/9/2025):
 - Cleaned up code files and comments
+- Changed gameover screen text
+- moved loadfont function to utils.lua